consolidate tags into package, deduplicate location and product tags
