--- api/src/main/kotlin/com/ford/internalprojects/peoplemover/location/LocationRequest.kt
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) 2020 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ford.internalprojects.peoplemover.location
-
-import javax.validation.constraints.*
-
-data class LocationRequest (
-    @field:NotBlank(message = "Name is required.")
-    @field:Size(max = 255, message = "Name must be less than 255 characters.")
-    val name: String
-)
-
-fun LocationRequest.toSpaceLocation(spaceUuid: String, id: Int) : SpaceLocation =
-        SpaceLocation(id = id, name = this.name, spaceUuid = spaceUuid)
