--- api/src/main/kotlin/com/ford/internalprojects/peoplemover/producttag/ProductTagRequest.kt
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2020 Ford Motor Company
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ford.internalprojects.peoplemover.producttag
-
-import javax.validation.constraints.NotBlank
-import javax.validation.constraints.Size
-
-data class ProductTagRequest (
-    @field:NotBlank(message = "Did you forget to provide a name for the product tag?")
-    @field:Size(max = 255, message = "Name cannot be longer than 255 characters")
-    val name: String
-)
